Added MessageSourceConfiguration and validation-messages.properties; Added DTO for Account, Product, Seller; Added mappers for DTO objects; Changed type for field price in Product from double to BigDecimal; Fixed application.yml. Changed property 'name' to 'username'; Added some integratin tests for CustomerRepositoryImplIntegrationTest, FeedbackRepositoryImplIntegrationTest, ProductRepositoryImplIntegrationTest, SellerRepositoryIntegrationTest, WishListRepositoryImplIntegrationTest; Added spring context test SimpleOnlineShopApplicationTest; Created controller for adding new products ProductController.class; Created tests for ProductController; Added home page index.html; Added page with form for adding new product add-new-product.html; Added page with result of created new product product-added.html;

+++ src/main/java/com/shop/online/simple/repository/rowmapper/ProductRowMapper.java
@@ -6,6 +6,7 @@ import com.shop.online.simple.entity.Product;
 import com.shop.online.simple.entity.Tag;
 import org.springframework.jdbc.core.ResultSetExtractor;
 
+import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.*;
@@ -65,7 +66,7 @@ public class ProductRowMapper implements ResultSetExtractor<List<Product>> {
         product.setId(id);
         product.setName(rs.getString("name"));
         product.setDescription(rs.getString("description"));
-        product.setPrice(rs.getDouble("price"));
+        product.setPrice(BigDecimal.valueOf(rs.getDouble("price")));
 
         return product;
     }
